% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{treetestdata}
\alias{treetestdata}
\title{ShinyTreeMenu Test Data}
\format{A data.table with 45 rows and 8 variables:
\describe{
  \item{levelX_id}{The id of the branch of that level. Note this must be unique to that branch / level}
  \item{levelX_name}{The name of the menu}
  ...
}}
\usage{
treetestdata
}
\description{
A dataset containing four levels of treedata intended show the
concept of how ShinyTreeMenu reads treedata.
}
\keyword{datasets}
