% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateShinyTreeMenu.R
\name{updateShinyTreeMenu}
\alias{updateShinyTreeMenu}
\title{Update the ShinyTreeMenu}
\usage{
updateShinyTreeMenu(treedata, treemenu_open, level_icons)
}
\arguments{
\item{treedata}{data.table; treedata - see details for more}

\item{treemenu_open}{reactive expression; Id belonging to the branch you want to open.}

\item{level_icons}{character; Names of fa-icons to use for the tree - see [ShinyTreeMenuHTML()] for more details.}
}
\value{
A HTML class character containing the tree menu
}
\description{
updateShinyTreeMenu() updates the tree menu, when a new level is selected
}
\examples{




# Create a reactive expression holding the treedata

treedata <- reactive({
   ShinyTreeMenu::treetestdata
})


# Wrap the renderShinyTreeMenu function in a shiny callModule

shiny::callModule(ShinyTreeMenu::renderShinyTreeMenu,
                  id = "treemenu",
                  treedata = treedata,
                  level_icons = "tree")


# To update your ShinyTreeMenu on selection use:

ShinyTreeMenu::updateShinyTreeMenu(
  treedata = treedata,
  treemenu_open = reactive({input$treemenu_open}),
  level_icons = "tree")


}
\seealso{
\code{\link{renderShinyTreeMenu}} and \code{\link{ShinyTreeMenuHTML}}
}
