% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderShinyTreeMenu.R
\name{renderShinyTreeMenu}
\alias{renderShinyTreeMenu}
\title{Render the ShinyTreeMenu HTML}
\usage{
renderShinyTreeMenu(input, output, session, treedata, level_icons)
}
\arguments{
\item{treedata}{data.table; treedata - see \code{\link{ShinyTreeMenuHTML}} for more details.}

\item{level_icons}{character; Names of fa-icons to use for the tree - see \code{\link{ShinyTreeMenuHTML}} for more details.}
}
\description{
This function should be run within the Shiny \code{\link[shiny]{callModule}}
function and renders a reactive ShinyTreeMenu HTML.
}
\details{
renderShinyTreeMenu is a dynamic HTML rendering function, that
runs within the Shiny \code{\link[shiny]{callModule}} function.

renderShinyTreeMenu renders the initial treemenu as a reactive
Shiny UI Output and this treemenu can then be updated using
\code{\link{updateShinyTreeMenu}}.
}
\examples{




# Create a reactive expression holding the treedata

treedata <- reactive({
   ShinyTreeMenu::treetestdata
})


# Wrap the renderShinyTreeMenu function in a shiny callModule

shiny::callModule(ShinyTreeMenu::renderShinyTreeMenu,
                  id = "treemenu",
                  treedata = treedata,
                  level_icons = "tree")


# To update your ShinyTreeMenu on selection use:

ShinyTreeMenu::updateShinyTreeMenu(
  treedata = treedata,
  treemenu_open = reactive({input$treemenu_open}),
  level_icons = "tree")


}
\seealso{
\code{\link{ShinyTreeMenuOutput}}, \code{\link{updateShinyTreeMenu}} and \code{\link{ShinyTreeMenuHTML}}
}
