% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyTreeMenuOutput.R
\name{ShinyTreeMenuOutput}
\alias{ShinyTreeMenuOutput}
\title{Render the ShinyTreeMenu HTML}
\usage{
ShinyTreeMenuOutput(outputId, class = NULL)
}
\arguments{
\item{outputId}{output variable to read the value from}

\item{class}{character; Additional CSS classes to add to the div output}
}
\description{
Render a reactive ShinyTreeMenu output variable in your Shiny UI.
}
\examples{

ShinyTreeMenuOutput("treemenu")

}
