% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateShinyTreeMenu.R
\name{updateShinyTreeMenu}
\alias{updateShinyTreeMenu}
\title{Update the ShinyTreeMenu}
\usage{
updateShinyTreeMenu(treedata, treemenu_open, level_icons)
}
\arguments{
\item{treedata}{data.table; treedata - see details for more}

\item{treemenu_open}{reactive expression; Id belonging to the branch you want to open.}

\item{level_icons}{character; Names of fa-icons to use for the tree - see [ShinyTreeMenuHTML()] for more details.}
}
\value{
A HTML class character containing the tree menu
}
\description{
updateShinyTreeMenu() updates the tree menu, when a new level is selected
}
\examples{


if (interactive()) {

  shinyApp(

    ui = fluidPage(

      # ShinyTreeMenu uses shinyjs - so remember to include shinyjs in your UI
      shinyjs::useShinyjs(),

      ShinyTreeMenuOutput("treemenu")
    ),

    server = function(input, output, session) {

      # Treedata has to be in the form of a reactive expression
      treedata <- reactive({
        ShinyTreeMenu::treetestdata
      })

      # Use the shiny callModule to render the ShinyTreeMenu
      # Note that you should not include () when adding your reactive treedata
      shiny::callModule(ShinyTreeMenu::renderShinyTreeMenu,
                        id = "treemenu",
                        treedata = treedata,
                        level_icons = "tree")

      # Here you will update the ShinyTreeMenu
      # As with the render function, you should not include () when adding your reactive treedata
      ShinyTreeMenu::updateShinyTreeMenu(
        treedata = treedata,
        treemenu_open = reactive({input$treemenu_open}),
        level_icons = "tree")


      # For the purpose of displaying the selection functionality
      observeEvent(input$treemenu, {
        showModal(
          modalDialog(
            paste("Selected id:", input$treemenu$val,
                  "| Selected level:", input$treemenu$level,
                  "| Selected row:", input$treemenu$row)
          )
        )
      })


    }

  )

}


}
\seealso{
\code{\link{ShinyTreeMenuOutput}}, \code{\link{renderShinyTreeMenu}} and \code{\link{ShinyTreeMenuHTML}}
}
